immortalwaifu/bot/requirements.txt
aiofiles==24.1.0
aiohttp==3.11.18
aiosignal==1.3.2
APScheduler==3.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
boto3==1.38.25
botocore==1.38.25
cachetools==4.2.2
certifi==2020.12.5
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
cloudscraper==1.2.71
decorator==5.2.1
dnspython==2.7.0
frozenlist==1.6.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==2.10
motor==3.7.0
multidict==6.4.4
ntgcalls==2.0.5
numpy==2.2.6
pillow==10.4.0
propcache==0.3.1
Pyrogram==2.0.106
pymongo==4.12.1
py-tgcalls
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
pytz==2025.2
requests==2.25.1
TgCrypto==1.2.5
typing_extensions==4.13.2
urllib3==1.26.5
yarl==1.20.0
yt-dlp==2025.5.22
